# web-app/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Second Brain API
  version: 1.0.0
  description: Session-authenticated JSON API over the indexed vault. Unsafe methods
    need the CSRF token header. The committed openapi.yaml is the contract (plan section
    5).
paths:
...
  /api/notes/:
    get:
      operationId: notes_list
      description: Filters combine with AND. Ordering defaults to `-modified`; every
        ordering ends with `path`. Paginated with `page` and `page_size` (default
        50, at most 200).
      summary: List notes
      parameters:
      - in: query
        name: due_after
        schema:
          type: string
          format: date
      - in: query
        name: due_before
        schema:
          type: string
          format: date
      - in: query
        name: has_parse_error
        schema:
          type: boolean
      - in: query
        name: ordering
        schema:
          enum:
          - -modified
          - due
          - title
          - path
          - -path
          type: string
          minLength: 1
        description: |-
          Default `-modified`. Every ordering ends with `path`.

          * `-modified` - -modified
          * `due` - due
          * `title` - title
          * `path` - path
          * `-path` - -path
      - in: query
        name: overdue
        schema:
          type: boolean
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
          minimum: 1
          maximum: 200
      - in: query
        name: path_prefix
        schema:
          type: string
          maxLength: 200
          minLength: 1
      - in: query
        name: priority
        schema:
          enum:
          - low
          - medium
          - high
          type: string
          minLength: 1
        description: |-
          * `low` - low
          * `medium` - medium
          * `high` - high
      - in: query
        name: project
        schema:
          type: string
          maxLength: 200
          minLength: 1
        description: Project slug.
      - in: query
        name: status
        schema:
          type: array
          items:
            type: string
        description: Repeatable.
      - in: query
        name: tag
        schema:
          type: string
          maxLength: 255
          minLength: 1
      - in: query
        name: type
        schema:
          type: string
          maxLength: 255
          minLength: 1
      tags:
      - notes
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNoteSummaryList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
    post:
      operationId: notes_create
      description: '`409` on a same-folder name collision; `422` for an unknown project.
        Implemented in P1-28.'
      summary: Create a note from a template
      tags:
      - notes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteSummary'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/notes/lookup/:
    get:
      operationId: notes_lookup
      description: Exactly one of `path` and `id`. An `id` matching several notes
        returns `409` with the candidate paths.
      summary: One note by path or id
      parameters:
      - in: query
        name: id
        schema:
          type: string
          maxLength: 255
          minLength: 1
        description: The `id` frontmatter value.
      - in: query
        name: path
        schema:
          type: string
          maxLength: 200
          minLength: 1
        description: Vault-relative path.
      tags:
      - notes
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmbiguousLookupError'
          description: ''
  /api/notes/status/:
    post:
      operationId: notes_status
      description: Validates the status against the vocabulary of the note's type.
        `409` on a hash mismatch, `422` on malformed frontmatter or a status outside
        the vocabulary. Implemented in P1-28.
      summary: Change a note's status
      tags:
      - notes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChangeRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDetail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
components:
  schemas:
...
    AmbiguousLookupError:
      type: object
      properties:
        detail:
          type: string
        candidates:
          type: array
          items:
            type: string
          description: Paths of the notes a `?id=` lookup matched.
      required:
      - candidates
      - detail
...
    Backlink:
      type: object
      properties:
        path:
          type: string
        title:
          type: string
      required:
      - path
      - title
...
    CreateNoteRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/TypeEnum'
        title:
          type: string
          maxLength: 200
        project:
          type: string
          description: Project slug or title.
          maxLength: 200
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        due:
          type: string
          format: date
        status:
          allOf:
          - $ref: '#/components/schemas/NoteStatusEnum'
          description: |-
            From the vocabulary of `type` (plan section 2.3); default per type.

            * `inbox` - inbox
            * `planned` - planned
            * `in-progress` - in-progress
            * `blocked` - blocked
            * `review` - review
            * `done` - done
            * `cancelled` - cancelled
            * `active` - active
            * `paused` - paused
            * `archived` - archived
            * `proposed` - proposed
            * `accepted` - accepted
            * `superseded` - superseded
            * `rejected` - rejected
            * `triaged` - triaged
            * `dismissed` - dismissed
        body:
          type: string
      required:
      - title
      - type
...
    Error:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
...
    NoteDetail:
      type: object
      description: One note in full.
      properties:
        id:
          type: string
          nullable: true
          description: The `id` frontmatter value; null for a note without one.
        path:
          type: string
          description: Vault-relative path.
        type:
          type: string
          description: A known type, `note`, or an unknown type as written.
        title:
          type: string
        status:
          type: string
          nullable: true
          description: Stored as written, even outside the type's vocabulary.
        priority:
          type: string
          nullable: true
          description: Stored as written; one of low/medium/high when valid.
        project:
          type: string
          nullable: true
          description: Project slug as written.
        due:
          type: string
          format: date
          nullable: true
          description: Null when absent or unreadable.
        blocked_by:
          type: string
          nullable: true
        decided:
          type: string
          format: date
          nullable: true
          description: 'Decisions: set when the status became accepted.'
        tags:
          type: array
          items:
            type: string
        created:
          type: string
          format: date
          nullable: true
        modified:
          type: string
          format: date-time
          description: Local time with the vault's offset.
        parse_error:
          type: string
          nullable: true
        frontmatter:
          type: object
          additionalProperties: {}
          description: The parsed frontmatter mapping.
        body:
          type: string
        content_hash:
          type: string
          description: Send back as `expected_hash` when writing.
        backlinks:
          type: array
          items:
            $ref: '#/components/schemas/Backlink'
          description: Notes that link here.
        links:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ResolvedLink'
          description: Each link target as written in the body, mapped to where it
            resolves.
      required:
      - backlinks
      - blocked_by
      - body
      - content_hash
      - created
      - decided
      - due
      - frontmatter
      - id
      - links
      - modified
      - parse_error
      - path
      - priority
      - project
      - status
      - tags
      - title
      - type
    NoteStatusEnum:
      enum:
      - inbox
      - planned
      - in-progress
      - blocked
      - review
      - done
      - cancelled
      - active
      - paused
      - archived
      - proposed
      - accepted
      - superseded
      - rejected
      - triaged
      - dismissed
      type: string
      description: |-
        * `inbox` - inbox
        * `planned` - planned
        * `in-progress` - in-progress
        * `blocked` - blocked
        * `review` - review
        * `done` - done
        * `cancelled` - cancelled
        * `active` - active
        * `paused` - paused
        * `archived` - archived
        * `proposed` - proposed
        * `accepted` - accepted
        * `superseded` - superseded
        * `rejected` - rejected
        * `triaged` - triaged
        * `dismissed` - dismissed
    NoteSummary:
      type: object
      description: Promoted fields of one indexed note.
      properties:
        id:
          type: string
          nullable: true
          description: The `id` frontmatter value; null for a note without one.
        path:
          type: string
          description: Vault-relative path.
        type:
          type: string
          description: A known type, `note`, or an unknown type as written.
        title:
          type: string
        status:
          type: string
          nullable: true
          description: Stored as written, even outside the type's vocabulary.
        priority:
          type: string
          nullable: true
          description: Stored as written; one of low/medium/high when valid.
        project:
          type: string
          nullable: true
          description: Project slug as written.
        due:
          type: string
          format: date
          nullable: true
          description: Null when absent or unreadable.
        blocked_by:
          type: string
          nullable: true
        decided:
          type: string
          format: date
          nullable: true
          description: 'Decisions: set when the status became accepted.'
        tags:
          type: array
          items:
            type: string
        created:
          type: string
          format: date
          nullable: true
        modified:
          type: string
          format: date-time
          description: Local time with the vault's offset.
        parse_error:
          type: string
          nullable: true
      required:
      - blocked_by
      - created
      - decided
      - due
      - id
      - modified
      - parse_error
      - path
      - priority
      - project
      - status
      - tags
      - title
      - type
    PaginatedNoteSummaryList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/NoteSummary'
...
    PriorityEnum:
      enum:
      - low
      - medium
      - high
      type: string
      description: |-
        * `low` - low
        * `medium` - medium
        * `high` - high
...
    ResolvedLink:
      type: object
      properties:
        path:
          type: string
          nullable: true
        state:
          $ref: '#/components/schemas/StateEnum'
      required:
      - path
      - state
...
    StateEnum:
      enum:
      - resolved
      - ambiguous
      - unresolved
      type: string
      description: |-
        * `resolved` - resolved
        * `ambiguous` - ambiguous
        * `unresolved` - unresolved
    StatusChangeRequest:
      type: object
      properties:
        path:
          type: string
          maxLength: 200
        status:
          allOf:
          - $ref: '#/components/schemas/NoteStatusEnum'
          description: |-
            Must belong to the vocabulary of the note's type; checked against the note.

            * `inbox` - inbox
            * `planned` - planned
            * `in-progress` - in-progress
            * `blocked` - blocked
            * `review` - review
            * `done` - done
            * `cancelled` - cancelled
            * `active` - active
            * `paused` - paused
            * `archived` - archived
            * `proposed` - proposed
            * `accepted` - accepted
            * `superseded` - superseded
            * `rejected` - rejected
            * `triaged` - triaged
            * `dismissed` - dismissed
        expected_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
        evidence:
          type: string
          description: Appended under `## Notes` when the status becomes `done`.
      required:
      - expected_hash
      - path
      - status
...
    TypeEnum:
      enum:
      - task
      - project
      - decision
      - lesson
      type: string
      description: |-
        * `task` - task
        * `project` - project
        * `decision` - decision
        * `lesson` - lesson
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid